requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2>=2.7.5,<2.8.0
flake8==3.9.2